+++ dejafu/CHANGELOG.rst
@@ -6,6 +6,26 @@ standard Haskell versioning scheme.
 
 .. _PVP: https://pvp.haskell.org/
 
+1.7.0.0 (2018-06-03)
+--------------------
+
+* Git: :tag:`dejafu-1.7.0.0`
+* Hackage: :hackage:`dejafu-1.7.0.0`
+
+Changed
+~~~~~~~
+
+* (:issue:`237`) ``Test.DejaFu.SCT.sctWeightedRandom`` and
+  ``sctWeightedRandomDiscard`` no longer take the number of executions
+  to use the same weights for as a parameter.
+
+Removed
+~~~~~~~
+
+* (:issue:`237`) The deprecated function
+  ``Test.DejaFu.Settings.swarmy``.
+
+
 1.6.0.0 (2018-05-11)
 --------------------